requirements.txt
Django==1.10.5
django-environ==0.4.1
django-ordered-model==1.3.0
djangorestframework==3.5.3
psycopg2==2.6.2
shortuuid==0.4.3
whitenoise==3.2.3
